backend/requirements.txt
# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Configuration and validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
PyJWT[crypto]>=2.8.0
httpx>=0.26.0
bcrypt>=4.0.0

# Image processing
pillow>=10.2.0
pillow-heif>=0.14.0

# Background jobs
redis>=5.0.1
arq>=0.25.0
watchfiles>=0.21.0  # Hot reload for worker in development

# Scheduling
apscheduler>=3.10.4

# Notifications
aiosmtplib>=3.0.1

# Utilities
tenacity>=8.2.3
python-dateutil>=2.8.2
imagehash>=4.3.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0
httpx>=0.26.0
